+++ eval.c
@@ -864,6 +864,19 @@ bool eval_xattr(const struct expr *expr, struct eval_state *state) {
 }
 
 /**
+ * -xattr test.
+ */
+bool eval_xattrname(const struct expr *expr, struct eval_state *state) {
+	int ret = bfs_check_xattr_named(state->ftwbuf, expr->sdata);
+	if (ret >= 0) {
+		return ret;
+	} else {
+		eval_report_error(state);
+		return false;
+	}
+}
+
+/**
  * -xtype test.
  */
 bool eval_xtype(const struct expr *expr, struct eval_state *state) {